made POST /admin/tutor_applications/:id/reject when invalid params are passed spec pass

+++ app/controllers/admin/tutor_applications_controller.rb
@@ -26,6 +26,11 @@ class Admin::TutorApplicationsController < ApplicationController
   def reject
     application = TutorApplication.find_by(id: params[:id])
 
+    if application.nil?
+      redirect_to admin_tutor_applications_pending_path, alert: "Invalid Learner was passed"
+      return
+    end
+
     application.destroy
     redirect_to admin_tutor_applications_pending_path, alert: "Application rejected"
   end
